chore: upgrade compression-webpack-plugin
This was needed because node-zopfli appears to be unmaintained and was causing issues with compressions-web-pack-plugin.
As a result, webpack needed to be upgraded to v3 due to the plugin depending on v3 of webpack.
Also as a result, extract-text-webpack-plugin and babel-loader also had to be upgraded to prevent any further warnings.